Child mutations are now deterministic. Now there can be more than 2 childs though it's extremely slow. Some refactoring.

+++ src/Logiikka/Otus.java
@@ -43,8 +43,8 @@ public class Otus {
     @Override
     public String toString() {
         StringBuilder builder = new StringBuilder();
-        for (Entry<Gene, Integer> gene : genomi) {
-            builder.append(gene.getValue());
+        for (Entry<Gene, Double> gene : genomi) {
+            builder.append(gene.getValue().intValue());
             builder.append(",");
         }
         if (builder.length() > 0) {
